First approach of estimated total cost implemented

+++ Business/IProductBusiness.cs
@@ -10,12 +10,12 @@ namespace BackendChallengeAPI.Business
     {
         public IEnumerable<Product> GetListOfAllProducts();
         public IEnumerable<Product> GetProductsByStatus(string status);
-        public IEnumerable<Product> GetProductsByRenewableRating(decimal rate);
+        public IEnumerable<Product> GetProductsByRenewableRating(double rate);
         public IEnumerable<Product> GetProductsByContractLength(int contractLength);
         public IEnumerable<Product> GetProductsBySupplier(string supplier);
         public bool AddProduct(Product product);
         public bool UpdateProduct(Product product);
         public bool DeleteProduct(int id);
-        public decimal GetProductEstimatedTotalCost(Product product, decimal estimatedConsumption);
+        public double GetProductEstimatedTotalCost(Product product, double estimatedConsumption);
     }
 }
